Add EMAX ES09MD servo demo program
Demo demonstrates how to work with digital servo
EMAX ES09MD without Arduino library.

Arduino Servo library is working slowly with this
fast digital servo.

PWM is used to control servo and voltage measurements
are used to calculate number of pulses to get the desired position.

Experimentally achieved constants for this servo
could be found in the servo.ino.

Tests:
* Min/max angle rotation test
* Equal steps test
* Random steps test